pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cinematch"
version = "0.1.0"
description = "Hybrid movie recommendation system combining content-based and collaborative filtering"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pgvector>=0.3",
    "pydantic-settings>=2.1",
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.8",
    "implicit>=0.7",
    "scipy>=1.12",
    "pandas>=2.2",
    "pyarrow>=15",
    "numpy>=1.26,<2.0",
    "redis[hiredis]>=5.0",
    "httpx>=0.27",
    "joblib>=1.3",
    "tqdm>=4.66",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
]

[project.optional-dependencies]
llm = ["llama-cpp-python>=0.2"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.9",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
